move async behind feature flags
